List session management functions in the API doc
Functions not exposed by a class, such as `getSessionFromStorage`, were
not listed as part of the public API. This introduces the same issue
that was identified in https://github.com/inrupt/solid-client-authn-js/pull/990
