+++ packages/node/tsconfig.json
@@ -8,7 +8,7 @@
 
   "typedocOptions": {
     "out": "website/docs/api/node",
-    "entryPoints": ["./src/index.ts"],
+    "entryPoints": ["./src/index.ts", "./src/multiSession.ts"],
     "entryDocument": "index.md"
   },
